Move Environment variables reducer
