+++ plugins/services/src/js/components/forms/EnvironmentFormSection.js
@@ -2,7 +2,7 @@ import {Tooltip} from 'reactjs-components';
 import {MountService} from 'foundation-ui';
 import React, {Component} from 'react';
 
-import {FormReducer as env} from '../../reducers/serviceForm/EnvironmentVariables';
+import {FormReducer as env} from '../../reducers/serviceForm/FormReducers/EnvironmentVariables';
 import {FormReducer as labels} from '../../reducers/serviceForm/FormReducers/Labels';
 import AddButton from '../../../../../../src/js/components/form/AddButton';
 import DeleteRowButton from '../../../../../../src/js/components/form/DeleteRowButton';
