+++ plugins/services/src/js/reducers/JSONMultiContainerParser.js
@@ -1,7 +1,7 @@
 import {JSONParser as container} from './serviceForm/Container';
 import {JSONParser as constraints} from './serviceForm/MultiContainerConstraints';
 import {JSONParser as fetch} from './serviceForm/JSONReducers/Artifacts';
-import {JSONParser as environmentVariables} from './serviceForm/EnvironmentVariables';
+import {JSONParser as environmentVariables} from './serviceForm/JSONReducers/EnvironmentVariables';
 import {JSONParser as externalVolumes} from './serviceForm/JSONReducers/ExternalVolumes';
 import {JSONParser as healthChecks} from './serviceForm/HealthChecks';
 import {JSONParser as labels} from './serviceForm/JSONReducers/Labels';
