+++ plugins/services/src/js/reducers/JSONMultiContainerReducers.js
@@ -1,6 +1,6 @@
 import {JSONReducer as constraints} from './serviceForm/MultiContainerConstraints';
 import {JSONReducer as containers} from './serviceForm/Containers';
-import {JSONReducer as env} from './serviceForm/EnvironmentVariables';
+import {JSONReducer as env} from './serviceForm/JSONReducers/EnvironmentVariables';
 import {JSONReducer as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONReducer as scaling} from './serviceForm/MultiContainerScaling';
 import {JSONReducer as labels} from './serviceForm/JSONReducers/Labels';
