+++ plugins/services/src/js/reducers/JSONSingleContainerReducers.js
@@ -1,6 +1,6 @@
 import {JSONReducer as constraints} from './serviceForm/JSONReducers/Constraints';
 import {JSONReducer as container} from './serviceForm/Container';
-import {JSONReducer as env} from './serviceForm/EnvironmentVariables';
+import {JSONReducer as env} from './serviceForm/JSONReducers/EnvironmentVariables';
 import {JSONReducer as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONReducer as healthChecks} from './serviceForm/HealthChecks';
 import {JSONReducer as labels} from './serviceForm/JSONReducers/Labels';
